move tests related metric to create-metrics.test.ts
