+++ src/main.test.ts
@@ -11,16 +11,10 @@ import {
 import { fetchWorkflowResults, WorkflowResults } from './github/index.js'
 import { run } from './main.js'
 import {
-  InMemoryMetricExporter,
-  AggregationTemporality,
-  MetricData
-} from '@opentelemetry/sdk-metrics'
-import {
   InMemorySpanExporter,
   ReadableSpan
 } from '@opentelemetry/sdk-trace-base'
 import * as instrumentation from './instrumentation/index.js'
-import { calcDiffSec } from './utils/calc-diff-sec.js'
 import { opentelemetryAllDisable } from './utils/opentelemetry-all-disable.js'
 
 const workflowRunResults = vi.hoisted(() => {
@@ -114,118 +108,6 @@ vi.mock(import('./github/index.js'), async importOriginal => {
 describe('run', () => {
   beforeEach(() => {
     opentelemetryAllDisable()
-  })
-
-  describe('should export expected metrics', () => {
-    let exporter: InMemoryMetricExporter
-    // TODO: testじゃなくてコード上で定義したい
-    const descriptorNames = {
-      TASK_DURATION: 'cicd.pipeline.task.duration',
-      TASK_QUEUED_DURATION: 'cicd.pipeline.task.queued_duration',
-      PIPELINE_DURATION: 'cicd.pipeline.duration',
-      PIPELINE_QUEUED_DURATION: 'cicd.pipeline.queued_duration'
-    }
-
-    beforeEach(() => {
-      // new instance because exporter.reset() is not work well.
-      exporter = new InMemoryMetricExporter(AggregationTemporality.DELTA)
-    })
-
-    test('should export expected descriptor name only', async () => {
-      await expect(run(exporter)).rejects.toThrow(
-        'process.exit unexpectedly called with "0"' // 0 is success
-      )
-      const expectedDescriptorNames = Object.values(descriptorNames)
-
-      const metrics = exporter.getMetrics()[0].scopeMetrics[0].metrics
-      for (const metric of metrics) {
-        expect(expectedDescriptorNames).toContain(metric.descriptor.name)
-      }
-    })
-
-    test('should verify cicd.pipeline.task.duration', async () => {
-      await expect(run(exporter)).rejects.toThrow(
-        'process.exit unexpectedly called with "0"' // 0 is success
-      )
-      const metric = findMetricByDescriptorName(
-        exporter,
-        descriptorNames.TASK_DURATION
-      )
-      const dataPoints = metric.dataPoints.map(dataPoint => ({
-        taskName: dataPoint.attributes['cicd.pipeline.task.name'],
-        value: dataPoint.value
-      }))
-
-      for (const job of workflowRunJobs) {
-        expect(dataPoints).toContainEqual({
-          taskName: job.name,
-          value: calcDiffSec(
-            new Date(job.started_at),
-            new Date(job.completed_at!)
-          )
-        })
-      }
-      expect(dataPoints).toHaveLength(workflowRunJobs.length)
-    })
-
-    test('should verify cicd.pipeline.task.queued_duration', async () => {
-      await expect(run(exporter)).rejects.toThrow(
-        'process.exit unexpectedly called with "0"' // 0 is success
-      )
-      const metric = findMetricByDescriptorName(
-        exporter,
-        descriptorNames.TASK_QUEUED_DURATION
-      )
-
-      const dataPoints = metric.dataPoints.map(dataPoint => ({
-        taskName: dataPoint.attributes['cicd.pipeline.task.name'],
-        value: dataPoint.value
-      }))
-
-      for (const job of workflowRunJobs) {
-        expect(dataPoints).toContainEqual({
-          taskName: job.name,
-          value: calcDiffSec(new Date(job.created_at), new Date(job.started_at))
-        })
-      }
-      expect(dataPoints).toHaveLength(workflowRunJobs.length)
-    })
-
-    test('should verify cicd.pipeline.duration', async () => {
-      await expect(run(exporter)).rejects.toThrow(
-        'process.exit unexpectedly called with "0"' // 0 is success
-      )
-      const metric = findMetricByDescriptorName(
-        exporter,
-        descriptorNames.PIPELINE_DURATION
-      )
-
-      expect(metric.dataPoints).toHaveLength(1)
-      expect(metric.dataPoints[0].value).toEqual(
-        calcDiffSec(
-          new Date(workflowRun.created_at),
-          new Date(workflowRunJobs[1].completed_at!) // last job's complete_at
-        )
-      )
-    })
-
-    test('should verify cicd.pipeline.queued_duration', async () => {
-      await expect(run(exporter)).rejects.toThrow(
-        'process.exit unexpectedly called with "0"' // 0 is success
-      )
-      const metric = findMetricByDescriptorName(
-        exporter,
-        descriptorNames.PIPELINE_QUEUED_DURATION
-      )
-
-      expect(metric.dataPoints).toHaveLength(1)
-      expect(metric.dataPoints[0].value).toEqual(
-        calcDiffSec(
-          new Date(workflowRun.created_at),
-          new Date(workflowRunJobs[0].started_at)
-        )
-      )
-    })
   })
 
   describe('should export expected spans', () => {
@@ -395,19 +277,6 @@ describe('run', () => {
     })
   })
 })
-
-const findMetricByDescriptorName = (
-  exporter: InMemoryMetricExporter,
-  name: string
-): MetricData => {
-  const metric = exporter
-    .getMetrics()[0]
-    .scopeMetrics[0].metrics.find(v => v.descriptor.name === name)
-  if (metric === undefined) {
-    throw new Error(`${name} descriptor is not found`)
-  }
-  return metric
-}
 
 const toEpochSec = (date: string): number => {
   return Math.floor(new Date(date).getTime() / 1000)
